pyproject.toml
[project]
name = "neocortex-ai"
version = "0.1.0"
description = "AI-powered developer skill analyzer & personalized learning assistant"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    # CLI
    "typer>=0.15.0",
    "rich>=13.0.0",

    # LLM Providers
    "anthropic>=0.52.0",
    "openai>=1.60.0",
    "google-genai>=1.0.0",

    # Content Fetching
    "httpx>=0.28.0",
    "readability-lxml>=0.8.0",
    "pymupdf>=1.25.0",

    # Utilities
    "tiktoken>=0.8.0",
    "cryptography>=44.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
neocortex = "neocortex.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neocortex"]
